# SPARKC-233 use executeAsync when joining with C* table