--- a/doc/reference.md
+++ b/doc/reference.md
@@ -196,6 +196,11 @@ OSS Cassandra this should never be used.</td>
   <td>Number of CQL rows fetched per driver request</td>
 </tr>
 <tr>
+  <td><code>input.join.throughput_query_per_sec</code></td>
+  <td>9223372036854775807</td>
+  <td>Maximum read throughput allowed per single core in query/s while joining RDD with C* table</td>
+</tr>
+<tr>
   <td><code>input.metrics</code></td>
   <td>true</td>
   <td>Sets whether to record connector specific metrics on write</td>