--- a/project/plugins.sbt
+++ b/project/plugins.sbt
@@ -20,3 +20,5 @@ addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.0.4")
 resolvers += "Spark Packages Main repo" at "https://dl.bintray.com/spark-packages/maven" 
 
 addSbtPlugin("org.spark-packages" %% "sbt-spark-package" % "0.2.3")
+
+addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.2.6")