This ought to speed up joining with C* table by using executeAsync
(instead of execute) and prematerializing C* ResultSets during join
operation. Concurrent querying is controlled by parameter
spark.cassandra.input.join.concurrent.reads
It specifies maximum number of ongoing executeAsync calls and thus
limits maximum number of prematerialized ResultSets.

Basically we buy some time for the cost of memory.

This also adds performance module with jmh plugin. Within module
there is a simple macrobenchmark that was used to generate following results:




before:

Benchmark |Mode | Cnt | Score | Error | Units
----------|-----|-----|-------|-------|------
measureOneElementPerPartitionsJoin | ss | 15 | 19252.805| ± 390.205| ms/op
measureSmallPartitionsJoin | ss | 15 | 20034.163| ± 1515.168| ms/op
measureModeratePartitionsJoin | ss | 15 | 20278.329| ± 1689.112| ms/op
measureBigPartitionsJoin | ss | 15 | 17515.329| ± 1420.479| ms/op


after:

Benchmark |Mode | Cnt | Score | Error | Units
----------|-----|-----|-------|-------|------
measureOneElementPerPartitionsJoin | ss | 15 | 8378.778 | ± 665.592 | ms/op
measureSmallPartitionsJoin | ss | 15 | 8468.695 | ± 891.526 | ms/op
measureModeratePartitionsJoin | ss | 15 | 8689.495 | ± 924.087 | ms/op
measureBigPartitionsJoin | ss | 15 | 8891.763 | ± 1208.863 | ms/op