@jtgrabowski 2016-04-21
@RussellSpitzer , below there is a comparison of original implementation and current (based on your suggestion). As one can imagine this is most beneficial for small and oneElement partitions.

Results look ok, indeed there is a speedup, but materializing this eagerly doesn't seem right. This solution is only ok when this is true:

> If the next operation is a shuffle the whole block will have to be materialized anyway.

I have not found the confirmation yet. I'll dig deeper but maybe you already know where to look and could point me to a right class?

Benchmark | Mode | Cnt | Score | Error |Units
----------|------|-----|-------|-------|-----
measureBigPartitionsJoin_old | ss | 10 | 1519.481 |± 254.452 | ms/op
measureBigPartitionsJoin_new | ss | 10 | 1351.666 |± 318.551 | ms/op
measureModeratePartitionsJoin_old | ss | 10 | 1830.509 |± 236.377 | ms/op
measureModeratePartitionsJoin_new | ss | 10 | 778.343 |± 194.493 | ms/op
measureSmallPartitionsJoin_old | ss | 10 | 9076.712 |± 200.409 | ms/op
measureSmallPartitionsJoin_new | ss | 10 | 1790.086 |± 205.705 | ms/op
measureOneElementPerPartitionsJoin_old | ss | 10 | 74703.597 |± 708.829 | ms/op
measureOneElementPerPartitionsJoin_new | ss | 10 | 12936.777 |± 384.309 | ms/op