@RussellSpitzer 2016-04-21
The materialization thing comes from spark need to serialize the entire partition before it can shuffle it. It doesn't move single elements from a partition during a shuffle. Since it's sending an entire block over the wire the whole partition must be in memory.  This is why you basically always need numCores * maxPartitionSize  Heap on your executors for reliable spark execution. You can sometimes spill during shuffles but it is sad times.

I wanted to up the throughput mostly because my mental model for C* is, 2ms a query but can handle roughly 20~30k queries per second. The only way to hit the throughput limit is to never block waiting for a response. Even at 5 queries in flight at a time we can only do 5 X 500 ~ 2500 (assuming perfect always 5 queries in flight) queries a second (given my imaginary latency). Anyway i'm mostly worried here about max throughput. 

If you look at the CassandraTableScanCode you'll see that it's actually materializing "fetchSize" rows at a time which is ~1000 by default. (guessing didn't check). So even when it looks like it's only doing one at a time the driver under the hood is pulling 1000 rows down.

Maybe we should just match the "fetchSize" input param. We'll obviously be underestimating for very wide partitions but it will be great for narrow partitions.