@jtgrabowski 2016-04-25
Thanks Russ.
I added a column with throughput called 'Queries/sec' for each entry in the latest results table. Each case has 200k rows to pull so the number of queries is derived from partitions size: big, moderate, small, oneElement which is (rows per partition): 5000, 500, 10, 1. Intuitively, this fix works better for cases with smaller partitions/bigger number of queries - for oneElement per partition we are around 15k/sec on my laptop.
It makes me think that we should give users a way to throttle this, WDYT?