@jtgrabowski 2016-04-26
So in other words, my proposal is to go with current implementation (materialize everything with one go) but give users the ability to slow down the process. `RateLimiter` class could nicely cut down the materialization process rate to desired queries/second value. The default could be not set.