@RussellSpitzer 2016-04-26
I like that idea, I think it's a good balance.