@jtgrabowski 2016-04-27
squashed and rebased on current master